utils: add socket source file
